% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{alignVRegions}
\alias{alignVRegions}
\title{alignVRegions}
\usage{
alignVRegions(sequences)
}
\arguments{
\item{sequences}{a DNAStringSet of sequences to align}
}
\value{
a list containing unique IDs for each group of exactly
aligned sequences in each v-region
}
\description{
For a group of sequences in RADlib, align the sequences in each
V-region individually.
}
\examples{
alignVRegions(sequences)
$V11
[1] "IW245_RS23890 taxid=619741 organism=\"Longispora fulva\" variable_region=1"
[2] "IW245_RS27830 taxid=619741 organism=\"Longispora fulva\" variable_region=1"
[3] "IW245_RS37080 taxid=619741 organism=\"Longispora fulva\" variable_region=1"
[4] "IW245_RS40075 taxid=619741 organism=\"Longispora fulva\" variable_region=1"
...
$V91
[1] "IW245_RS23890 taxid=619741 organism=\"Longispora fulva\" variable_region=9"
[2] "IW245_RS27830 taxid=619741 organism=\"Longispora fulva\" variable_region=9"
[3] "IW245_RS37080 taxid=619741 organism=\"Longispora fulva\" variable_region=9"
[4] "IW245_RS40075 taxid=619741 organism=\"Longispora fulva\" variable_region=9"
}
