% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createRADq}
\alias{createRADq}
\title{createRADq}
\usage{
createRADq(species_list, return_dataframe = FALSE)
}
\arguments{
\item{return_dataframe}{a boolean indicating whether a dataframe
containing the summary data should be returned in addition to the
csv created by default.}

\item{sequences}{a vector containing species names}
}
\value{
a dataframe containing the summary data when return_dataframe = TRUE
}
\description{
Given a list of species names, pull sequences for each species
from RADlibV, align them, and return either a csv or csv and
dataframe containing a summary of which variable regions aligned
exactly, designated by a unique ID for each group.
}
\examples{
createRADq(c("Pseudomonas aeruginosa"), TRUE)
                       species variable_region    copy_id seq_id
1  Pseudomonas aeruginosa PAO1              V1   PA0668.1    V11
2  Pseudomonas aeruginosa PAO1              V1   PA4280.5    V11
...
62 Pseudomonas aeruginosa PAO1              V9   PA4690.5    V92
63 Pseudomonas aeruginosa PAO1              V9   PA5369.5    V92
}
