% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createRADqGroups}
\alias{createRADqGroups}
\title{createRADqGroups}
\usage{
createRADqGroups(vregions, return_df = FALSE)
}
\arguments{
\item{vregions}{a vector containing all variable regions to be used
to split the groups}

\item{return_df}{a boolean indicating whether a dataframe
containing the data should be returned in addition to the
csv created by default.}
}
\value{
a dataframe containing the summary data when return_dataframe = TRUE
}
\description{
After createSummarizedIDs has been run, combines all summarized IDs
and sorts taxa into groups that share all the same IDs for the given
variable regions.
}
\examples{
createRADqGroups(c("V4","V5"), TRUE)
                    taxa    groups
1          Brucella suis    V41V51
2 Pseudomonas aeruginosa V42V52V53
}
