% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createSummarizedIDs}
\alias{createSummarizedIDs}
\title{createSummarizedIDs}
\usage{
createSummarizedIDs(return_df = FALSE)
}
\arguments{
\item{return_df}{a boolean indicating whether a dataframe
containing the data should be returned in addition to the
csv created by default.}
}
\value{
a dataframe containing the summary data when return_dataframe = TRUE
}
\description{
After createRADq has been run, combines all unique IDs for each v-region in
each species into a single ID.
}
\examples{
createSummarizedIDs(TRUE)
                 species  V1  V2  V3  V4     V5  V6  V7  V8  V9
1 Pseudomonas aeruginosa V11 V21 V31 V41 V51V52 V61 V71 V81 V91
}
