% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createSummary}
\alias{createSummary}
\title{createSummary}
\usage{
createSummary(IDs, return_df = FALSE)
}
\arguments{
\item{IDs}{a list containing unique IDs for each group of exactly
aligned sequences in each v-region}

\item{return_df}{a boolean indicating whether a dataframe
containing the summary data should be returned in addition to the
csv created by default.}
}
\value{
a dataframe containing the summary data when return_df = TRUE
}
\description{
Takes the list of IDs created by alignVRegions and summarizes the
data in a csv. Can also return the summary as a dataframe, if
return_df is set to true.
}
\examples{
createSummary(IDs, TRUE)
                       species variable_region    copy_id seq_id
1  Pseudomonas aeruginosa PAO1              V1   PA0668.1    V11
2  Pseudomonas aeruginosa PAO1              V1   PA4280.5    V11
...
62 Pseudomonas aeruginosa PAO1              V9   PA4690.5    V92
63 Pseudomonas aeruginosa PAO1              V9   PA5369.5    V92
}
