% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_to_organisms.R
\name{get_accessions_df}
\alias{get_accessions_df}
\title{get_accessions_df}
\usage{
get_accessions_df()
}
\value{
This function returns a callable df with two columns: species_name and accession_id
}
\description{
This function generates a callable df with two columns: species_name and accession_id
}
\examples{
get_accessions_df()
> species_name accession_id
}
