% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_organism_names}
\alias{get_organism_names}
\title{get_organism_names}
\usage{
get_organism_names(acc_ids)
}
\arguments{
\item{acc_ids}{\if{html}{\out{<char list>}} list of valid accession ids}
}
\value{
\if{html}{\out{<char list>}} list of organism names (includes duplicates)
}
\description{
Given a list of accession ids, outputs corresponding organism names (includes duplicates).
}
\examples{
> get_organism_names(c("EDX97_RS04345", "EDX97_RS05225", "EDX97_RS06840",
"EDX97_RS09045", "EDX97_RS10020", "EDX97_RS11935", "MOZ64_RS11590"))
[1] "Absicoccus porci"        "Absicoccus porci"        "Absicoccus porci"
[4] "Absicoccus porci"       "Absicoccus porci"        "Absicoccus porci"
[7] "Absicoccus intestinalis"
}
