% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_species_from_genus}
\alias{get_species_from_genus}
\title{get_species_from_genus}
\usage{
get_species_from_genus(genus_label)
}
\arguments{
\item{genus_label}{\if{html}{\out{<char>}} genus label with format "Genus - All Species (#)"}
}
\value{
\if{html}{\out{<char list>}} list of organism names
}
\description{
Finds all organisms in RADlib of a given genus.
}
\examples{
> get_species_from_genus("Absicoccus - All Species (2)")
[1] "Absicoccus porci"        "Absicoccus intestinalis"
}
