% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_species_list}
\alias{get_species_list}
\title{get_species_list}
\usage{
get_species_list(ids)
}
\arguments{
\item{accession_ids}{\if{html}{\out{<char list>}} of valid accession ids}
}
\value{
\if{html}{\out{<char list>}} of organism names (includes duplicates)
}
\description{
Given a list of accession ids, outputs corresponding organism names (includes duplicates)
}
\examples{
> get_species_list(c("EDX97_RS04345", "EDX97_RS05225", "EDX97_RS06840",
"EDX97_RS09045", "EDX97_RS10020", "EDX97_RS11935", "MOZ64_RS11590"))
[1] "Absicoccus porci"        "Absicoccus porci"        "Absicoccus porci"
[4] "Absicoccus porci"       "Absicoccus porci"        "Absicoccus porci"
[7] "Absicoccus intestinalis"
}
