% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_taxa_ids}
\alias{get_taxa_ids}
\title{get_taxa_ids}
\usage{
get_taxa_ids(acc_ids)
}
\arguments{
\item{acc_ids}{\if{html}{\out{<char list>}} list of valid accession ids}
}
\value{
\if{html}{\out{<char list>}} list of taxa ids (includes duplicates)
}
\description{
Given a list of accession ids, outputs corresponding taxa ids (includes duplicates).
}
\examples{
> get_taxa_ids(c("EDX97_RS04345", "EDX97_RS05225", "EDX97_RS06840",
"EDX97_RS09045", "EDX97_RS10020", "EDX97_RS11935", "MOZ64_RS11590"))
[1] "2486576" "2486576" "2486576" "2486576" "2486576" "2486576" "2926319"
}
