% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{getSequences}
\alias{getSequences}
\title{getSequences}
\usage{
getSequences(taxa)
}
\arguments{
\item{taxa}{a vector containing species names}
}
\value{
a list of DNAStringSet objects containing the sequences for
each variable region for each species.
}
\description{
Given a list of species, retrieves all sequences associated with
those taxa from RADlibV.
}
\examples{
getSequences(c("Longispora fulva"))
DNAStringSet object of length 36:
     width seq                                              names
 [1]    21 GAAAGGCCCTTCGGGGTACTC                            IW245_RS23890 tax...
 [2]   107 CTTGGCTTCGGGATAACCATCGG...GCCAGGGATGGGCTCGCGGCCT IW245_RS23890 tax...
 ...   ... ...
[36]    33 GCCGGTGGCCCAACCCGTAAGGGAGGGAGCCGT                IW245_RS40075 tax...
}
