% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADsequences.R
\name{getVRegions}
\alias{getVRegions}
\title{getVRegions}
\usage{
getVRegions(sequences, vregions)
}
\arguments{
\item{sequences}{\url{Biostrings::DNAStringSet} set of sequences to search}

\item{vregions}{\if{html}{\out{<char list>}} list of desired v-regions}
}
\value{
\url{Biostrings::DNAStringSet} subset of the input DNAStringSet containing all sequences from
specificed v-regions
}
\description{
Given a DNAStringSet of sequences with V-regions included in their names,
retrieves all instances of selected V-regions from the DNAStringSet.
}
\examples{
> getVRegions(my_string_set, c("V1", "V2"))

}
