% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADsequences.R
\name{readSequences}
\alias{readSequences}
\title{readSequences}
\usage{
readSequences(infile, accessions)
}
\arguments{
\item{infile}{\if{html}{\out{<char>}} file path to the desired library}

\item{accessions}{\if{html}{\out{<char list>}} list of accession ids}
}
\value{
\url{Biostrings::DNAStringSet} set of sequences
}
\description{
Reads in all sequences that include the given accession ids in their header.
}
\examples{
> readSequences("/my/file/path", c("1234.1", "1234.2", "1234.3"))

}
