% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{alignVRegions}
\alias{alignVRegions}
\title{alignVRegions}
\usage{
alignVRegions(sequences)
}
\arguments{
\item{sequences}{a DNAStringSet of sequences to align}
}
\value{
a list containing unique IDs for each group of exactly
aligned sequences in each v-region
}
\description{
For a group of sequences in RADlib, align the sequences in each
V-region individually
}
\examples{
alignVRegions(sequences)
$V11
[1] "GCF_000006765.1.1__V1" "GCF_000006765.1.2__V1" "GCF_000006765.1.3__V1"
[4] "GCF_000006765.1.4__V1"
...
$V91
[1] "GCF_000006765.1.1__V9" "GCF_000006765.1.2__V9" "GCF_000006765.1.3__V9"
[4] "GCF_000006765.1.4__V9"
}
