% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{alignVRegions}
\alias{alignVRegions}
\title{alignVRegions}
\usage{
alignVRegions(sequences)
}
\arguments{
\item{sequences}{a DNAStringSet of sequences to align}
}
\value{
a dataframe containing information about which sequences align exactly
}
\description{
For a group of sequences in RADlib, align the sequences in each V-region individually
}
\examples{
## TODO::still in progress
alignVRegions(sequences)
}
