% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createRADq}
\alias{createRADq}
\title{createRADq}
\usage{
createRADq(species_list, return_dataframe = FALSE)
}
\arguments{
\item{return_dataframe}{a boolean indicating whether a dataframe
containing the summary data should be returned in addition to the
csv created by default.}

\item{sequences}{a vector containing species names}
}
\value{
a dataframe containing the summary data when return_dataframe = TRUE
}
\description{
Given a list of species names, pull sequences for each species
from RADlibV, align them, and return either a csv or csv and
dataframe containing a summary of which variable regions aligned
exactly, designated by a unique ID for each group.
}
\examples{
createRADq(c("Pseudomonas aeruginosa"), TRUE)
createSummary(IDs, TRUE)
                  species variable_region copy_num seq_id
1  Pseudomonas aeruginosa              V1        1    V11
2  Pseudomonas aeruginosa              V1        2    V11
...
35 Pseudomonas aeruginosa              V9        3    V91
36 Pseudomonas aeruginosa              V9        4    V91
}
