% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createRADq}
\alias{createRADq}
\title{createRADq}
\usage{
createRADq(species_list, return_dataframe = FALSE)
}
\arguments{
\item{return_dataframe}{a boolean indicating whether a dataframe
containing the summary data should be returned in addition to the
csv created by default.}

\item{sequences}{a vector containing species names}
}
\value{
a dataframe containing the summary data when return_dataframe = TRUE
}
\description{
Given a list of species names, pull sequences for each species
from RADlibV, align them, and return either a csv or csv and
dataframe containing a summary of which variable regions aligned
exactly, designated by a unique ID for each group.
}
\examples{
## TODO::still in progress
alignVRegions(sequences)
$V11
[1] "GCF_000006765.1.1__V1" "GCF_000006765.1.2__V1" "GCF_000006765.1.3__V1"
[4] "GCF_000006765.1.4__V1"
...
$V91
[1] "GCF_000006765.1.1__V9" "GCF_000006765.1.2__V9" "GCF_000006765.1.3__V9"
[4] "GCF_000006765.1.4__V9"
}
