% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{createSummary}
\alias{createSummary}
\title{createSummary}
\usage{
createSummary(IDs, return_df = FALSE)
}
\arguments{
\item{IDs}{a list containing unique IDs for each group of exactly
aligned sequences in each v-region}

\item{return_df}{a boolean indicating whether a dataframe
containing the summary data should be returned in addition to the
csv created by default.}
}
\value{
a dataframe containing the summary data when return_df = TRUE
}
\description{
Takes the list of IDs created by alignVRegions and summarizes the
data in a csv. Can also return the summary as a dataframe, if
return_df is set to true.
}
\examples{
createSummary(IDs, TRUE)
                  species variable_region copy_num seq_id
1  Pseudomonas aeruginosa              V1        1    V11
2  Pseudomonas aeruginosa              V1        2    V11
...
35 Pseudomonas aeruginosa              V9        3    V91
36 Pseudomonas aeruginosa              V9        4    V91
}
