% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADdownload.R
\name{download_MetaScope_reference}
\alias{download_MetaScope_reference}
\title{download_MetaScope_reference}
\usage{
download_MetaScope_reference(
  accessions_list,
  download_folder,
  filter = FALSE,
  ref_library = "RADlib"
)
}
\arguments{
\item{accessions_list}{\if{html}{\out{<char list>}} list of accession ids of sequences to download from RADlib}

\item{download_folder}{\if{html}{\out{<char>}} path to desired destination folder (sub folder will be created)}

\item{filter}{\if{html}{\out{<boolean>}} (optional) for filter/reference sequence download, defaults to FALSE}
}
\value{
\if{html}{\out{<list>}} with \if{html}{\out{<char>}} $folder containing the sub folder name and \if{html}{\out{<char list>}} $files containing a list of downloaded fasta file names
}
\description{
Downloads selected sequences from RADlib to a desired folder as fasta files and places them inside a single sub folder.
}
\examples{
> download_MetaScope_reference(c("EDX97_RS10020", "EDX97_RS11935", "MOZ64_RS11590"), "/Users/user/Downloads")
$folder
[1] "MetaScope_reference_dir"
$files
$files[[1]]
[1] "Absicoccus_porci_EDX97_RS10020.fasta"
$files[[2]]
[1] "Absicoccus_porci_EDX97_RS11935.fasta"
$files[[3]]
[1] "Absicoccus_intestinalis_MOZ64_RS11590.fasta"
}
