% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADdownload.R
\name{download_MetaScope_reference}
\alias{download_MetaScope_reference}
\title{download_MetaScope_reference}
\usage{
download_MetaScope_reference(accessions_list, download_folder, filter = FALSE)
}
\arguments{
\item{accessions_list}{list of string accession ids of sequences to download from RADlib.}

\item{download_folder}{character path to desired destination folder}
}
\value{
This function downloads a filtered portion of RADlib (fasta) and outputs the file name.
}
\description{
This function filters RADlib by accession ids and downloads a reference database to the provided folder
}
\examples{
download_MetaScope_reference(c("NZ_CTYB01000002.1","NZ_CTYB01000003.1"), /Users/user/Downloads/RAD_downloads_folder)
> "Metascope_reference_db.fasta"
}
