% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADdownload.R
\name{download_RAD_data}
\alias{download_RAD_data}
\title{download_RAD_data}
\usage{
download_RAD_data(
  pipeline,
  organisms_list,
  filter = FALSE,
  download_location = fs::path_home("Downloads")
)
}
\arguments{
\item{pipeline}{\if{html}{\out{<char>}} name of pipeline, valid inputs: "MetaScope"}

\item{organisms_list}{\if{html}{\out{<char list>}} of organism names to download from RADlib}

\item{filter}{\if{html}{\out{<boolean>}} (optional) for filter/reference sequence download, defaults to FALSE}

\item{download_location}{\if{html}{\out{<char>}} (optional) file path to desired download location, defaults to "Downloads" folder}
}
\value{
\if{html}{\out{<char>}} path to downloaded folder of files
}
\description{
This function downloads files for integration with analysis pipelines, currently supports MetaScope.
}
\examples{
> download_RAD_data("MetaScope", c("Pseudomonas aeruginosa", "Brucella suis"))
"Users/user/Downloads/RADdownloads_05032026_204428_QVrV4idv/MetaScope_reference_dir"
}
