% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_RAD_data.R
\name{download_RAD_data}
\alias{download_RAD_data}
\title{download_RAD}
\usage{
download_RAD_data(pipeline, species_list, download_location = getwd())
}
\arguments{
\item{pipeline}{name of pipeline. Valid inputs: ("MetaScope", "Kraken")}

\item{species_list}{list of species names to download from RADlib.}

\item{download_location}{optional parameter, file path for where downloaded files should go, defaults to working directory}
}
\value{
This function downloads a zipped folder to the user's working directory, and outputs the location of the downloaded folder.
}
\description{
This function allows users to download files for integration with analysis pipelines, currently supports MetaScope and Kraken.
}
\examples{
download_RADsynch("MetaScope", c("Pseudomonas aeruginosa", "Brucella suis"))
> "Downloaded Successfully to: ~/user/Downloads"
}
