% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_to_organisms.R
\name{get_accession_ids}
\alias{get_accession_ids}
\title{get_accession_ids}
\usage{
get_accession_ids(organisms)
}
\arguments{
\item{species_list}{list of valid species names}
}
\value{
list of accession ids (with copy numbers)
}
\description{
Given a list of species, outputs accession ids
}
\examples{
get_accession_ids(c("GCF_000006765.1.1", "GCF_000006765.1.2",
"GCF_000006765.1.3", "GCF_000006765.1.4", "GCF_000007505.1.1",
"GCF_000007505.1.2", "GCF_000007505.1.3"))
[1] "Pseudomonas aeruginosa" "Pseudomonas aeruginosa" "Pseudomonas aeruginosa" "Pseudomonas aeruginosa"
[5] "Brucella suis"          "Brucella suis"          "Brucella suis"
}
