% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_accession_ids}
\alias{get_accession_ids}
\title{get_accession_ids}
\usage{
get_accession_ids(organisms)
}
\arguments{
\item{organisms}{\if{html}{\out{<char list>}} of valid organism names}
}
\value{
\if{html}{\out{<char list>}} of accession ids
}
\description{
Given a list of organism names, outputs all corresponding accession ids
}
\examples{
> get_accession_ids(c("Absicoccus porci", "Absicoccus intestinalis"))
"GCF_000007505.1.2", "GCF_000007505.1.3"))
[1] "EDX97_RS04345" "EDX97_RS05225" "EDX97_RS06840" "EDX97_RS09045"
[5] "EDX97_RS10020" "EDX97_RS11935" "MOZ64_RS11590"
}
