% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accession_ids.R
\name{get_accession_ids}
\alias{get_accession_ids}
\title{get_accession_ids}
\usage{
get_accession_ids(species_list)
}
\arguments{
\item{species_list}{list of valid species names}
}
\value{
list of accession ids (with copy numbers)
}
\description{
Given a list of species, outputs accession ids
}
\examples{
get_accession_ids(c("Pseudomonas aeruginosa", "Brucella suis"))
[1] "GCF_000006765.1.1" "GCF_000006765.1.2" "GCF_000006765.1.3" "GCF_000006765.1.4"
[5] "GCF_000007505.1.1" "GCF_000007505.1.2" "GCF_000007505.1.3"
}
