% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_accessions_df}
\alias{get_accessions_df}
\title{get_accessions_df}
\usage{
get_accessions_df()
}
\value{
<data.frame> table with \if{html}{\out{<char>}} columns: id, taxid, organism, and genus.
}
\description{
Generates a metadata table from the header data of RADlib.
}
\examples{
> get_accessions_df()
               id   taxid               organism               genus
           <char>  <char>                 <char>              <char>
1: AB6B37_RS01935  570162    Fretibacter rubidus         Fretibacter
2:  KK102_RS15670 2838947 Enterobacter quasimori        Enterobacter
3:  KK102_RS22065 2838947 Enterobacter quasimori        Enterobacter
}
