% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_to_organisms.R
\name{get_all_organisms}
\alias{get_all_organisms}
\title{get_all_organisms}
\usage{
get_all_organisms()
}
\value{
list of organism names
}
\description{
outputs all unique organisms in RADlib
}
\examples{
> head(get_all_organisms())
[1] ""                                                     " - All Species"
[3] "'Nostoc - All Species"                                "'Nostoc azollae' 0708"
[5] "[Acidovorax] - All Species"                           "[Acidovorax] ebreus TPSY"
}
