% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADaccessions.R
\name{get_all_organisms}
\alias{get_all_organisms}
\title{get_all_organisms}
\usage{
get_all_organisms()
}
\value{
\if{html}{\out{<char list>}} of organism names (and genus labels)
}
\description{
Generates a sorted list of all unique organisms in RADlib and adds genus labels ("Genus - All Species (#)").
}
\examples{
> get_all_organisms()
[1] "Abditibacterium - All Species (1)" "Abditibacterium utsteinense"       "Abiotrophia - All Species (1)"
[4] "Abiotrophia defectiva"             "Absicoccus - All Species (2)"      "Absicoccus intestinalis"
}
