% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_to_organisms.R
\name{get_species_from_genus}
\alias{get_species_from_genus}
\title{get_species_from_genus}
\usage{
get_species_from_genus(genus_label)
}
\value{
list of organism names
}
\description{
outputs all organisms in RADlib of a given genus
}
\examples{
> get_species_from_genus("Segatella - All Species")
[1] "Segatella hominis"             "Segatella oris"                "Segatella copri DSM 18205"     "Segatella baroniae DSM 16972 "
[5] "Segatella cerevisiae"          "Segatella maculosa OT 289"     "Segatella bryantii"            "Segatella asaccharophila"
[9] "Segatella intestinalis"        "Segatella oulorum F0390"       "Segatella salivae F0493"       "Segatella sinensis"
}
