% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_to_organisms.R
\name{get_species_list}
\alias{get_species_list}
\title{get_species_list}
\usage{
get_species_list(ids)
}
\arguments{
\item{accession_ids}{list of valid accession ids}
}
\value{
list of species (with duplicates)
}
\description{
Given a list of accession ids, outputs species list (with duplicates)
}
\examples{
get_accession_ids(c("Pseudomonas aeruginosa", "Brucella suis"))
[1] "GCF_000006765.1.1" "GCF_000006765.1.2" "GCF_000006765.1.3" "GCF_000006765.1.4"
[5] "GCF_000007505.1.1" "GCF_000007505.1.2" "GCF_000007505.1.3"
}
