% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{getSequences}
\alias{getSequences}
\title{getSequences}
\usage{
getSequences(taxa)
}
\arguments{
\item{taxa}{a vector containing species names}
}
\value{
a list of DNAStringSet objects containing the sequences for
each variable region for each species.
}
\description{
Given a list of species, retrieves all sequences associated with
those taxa from RADlibV.
}
\examples{
createRADq(c("Pseudomonas aeruginosa"), TRUE)
                  species variable_region copy_num seq_id
1  Pseudomonas aeruginosa              V1        1    V11
2  Pseudomonas aeruginosa              V1        2    V11
...
35 Pseudomonas aeruginosa              V9        3    V91
36 Pseudomonas aeruginosa              V9        4    V91
}
