% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADlibHandler.r
\name{getVRegions}
\alias{getVRegions}
\title{getVRegions}
\usage{
getVRegions(sequences, vregions)
}
\arguments{
\item{sequences}{a DNAStringSet of sequences to search}

\item{vregions}{a list of v-regions to search for in the DNAStringSet}
}
\value{
a subset of the input DNAStringSet containing all sequences from
specificed v-regions.
}
\description{
Given a DNAStringSet of sequences with V-regions included in their names,
retrieve all instances of selected V-regions from the DNAStringSet.
}
\examples{
## getVRegions(my_string_set, c("V1", "V2"))
}
