% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADlibHandler.R
\name{readSequences}
\alias{readSequences}
\title{readSequences}
\usage{
readSequences(infile, accessions)
}
\arguments{
\item{infile}{The file path to the desired library}

\item{accessions}{A list of accession numbers to read in}
}
\value{
A Biostrings DNAStringSet containing all sequences that had one
of the given accession numbers in their identification line
}
\description{
Read in all sequences that include the given accession numbers in their
identification line.
}
\examples{
readSequences("/my/file/path", c("1234.1", "1234.2", "1234.3"))
}
