% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADalign.R
\name{selectVRegions}
\alias{selectVRegions}
\title{selectVRegions}
\usage{
selectVRegions(vregions, return_df = FALSE)
}
\arguments{
\item{vregions}{a vector of variable regions to include in the
filtered file}

\item{return_df}{a boolean indicating whether a dataframe
containing the data should be returned in addition to the
csv created by default.}
}
\value{
a dataframe containing the summary data when return_dataframe = TRUE
}
\description{
After createRADq has been run, filters the csv file from
createRADq to include only user-specificed variable regions.
}
\examples{
selectVRegions(c("V1", "V5"))
   X                species variable_region copy_num seq_id
}
